reorganize the samples.
