+++ AzureAIContentSafety.sln
@@ -21,6 +21,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AnalyzeText", "dotnet\1.0.0
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManageBlocklist", "dotnet\1.0.0\ManageBlocklist\ManageBlocklist.csproj", "{946B8E9C-E540-4C83-8E91-B9AEDAFE9759}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AnalyzeTextWithBlocklist", "AnalyzeTextWithBlocklist\AnalyzeTextWithBlocklist.csproj", "{92B2364E-1287-486F-A82C-3A8CF9CDE46F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OtherBlocklistOperations", "OtherBlocklistOperations\OtherBlocklistOperations.csproj", "{4A906BA7-C40E-4823-A7D3-37B9D2704976}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +55,14 @@ Global
 		{946B8E9C-E540-4C83-8E91-B9AEDAFE9759}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{946B8E9C-E540-4C83-8E91-B9AEDAFE9759}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{946B8E9C-E540-4C83-8E91-B9AEDAFE9759}.Release|Any CPU.Build.0 = Release|Any CPU
+		{92B2364E-1287-486F-A82C-3A8CF9CDE46F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{92B2364E-1287-486F-A82C-3A8CF9CDE46F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{92B2364E-1287-486F-A82C-3A8CF9CDE46F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{92B2364E-1287-486F-A82C-3A8CF9CDE46F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4A906BA7-C40E-4823-A7D3-37B9D2704976}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4A906BA7-C40E-4823-A7D3-37B9D2704976}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4A906BA7-C40E-4823-A7D3-37B9D2704976}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4A906BA7-C40E-4823-A7D3-37B9D2704976}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -64,6 +76,8 @@ Global
 		{E5BBA5F2-1D9C-42E9-9779-DD1A455E01EF} = {66A8E1A8-D9DD-4D36-B80A-C3EA8B7E58CC}
 		{5BE19581-5F42-4F8A-A6A6-A1093A2900D3} = {66A8E1A8-D9DD-4D36-B80A-C3EA8B7E58CC}
 		{946B8E9C-E540-4C83-8E91-B9AEDAFE9759} = {66A8E1A8-D9DD-4D36-B80A-C3EA8B7E58CC}
+		{92B2364E-1287-486F-A82C-3A8CF9CDE46F} = {66A8E1A8-D9DD-4D36-B80A-C3EA8B7E58CC}
+		{4A906BA7-C40E-4823-A7D3-37B9D2704976} = {66A8E1A8-D9DD-4D36-B80A-C3EA8B7E58CC}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {DE6CD92F-7A11-4462-9F80-534EBC54884B}
